Fixed bugs with code on paramedic side
At first it was assumed that data in firebase wasn't meant to be modified with the use of fragments. This was proven to be incorrect.
This assumption came about due to limited resources available online with regards to use of database with fragments, and also due to our limited understanding of the inner-workings of firebase libraries.

Cause of bugs for the paramedic side:  firebase code wasn't utilized properly